Modularizing logic. Templating.
