+++ src/midgard/CMakeLists.txt
@@ -2,6 +2,7 @@ file(GLOB headers ${VALHALLA_SOURCE_DIR}/valhalla/midgard/*.h)
 
 set_source_files_properties(
         polyline2.cc
+        point_tile_index.cc
         PROPERTIES COMPILE_FLAGS "-Werror -Wuninitialized")
 
 set(sources
@@ -10,6 +11,7 @@ set(sources
   polyline2.cc
   obb2.cc
   pointll.cc
+  point_tile_index.cc
   aabb2.cc
   point2.cc
   util.cc
